Moved global colors to 'styles' subdir
