+++ styles/Colors.ts
@@ -1,0 +1,24 @@
+const tintColorLight = "#2f95dc";
+const tintColorDark = "#fff";
+
+export default {
+	light: {
+		text: "#000",
+		background: "#fff",
+		tint: tintColorLight,
+		tabIconDefault: "#ccc",
+		tabIconSelected: tintColorLight,
+	},
+	dark: {
+		text: "#fff",
+		background: "#000",
+		tint: tintColorDark,
+		tabIconDefault: "#ccc",
+		tabIconSelected: tintColorDark,
+	},
+	uva: {
+		orange: "#E57202",
+		blue: "#002F6C",
+		white: "#FEFEFE",
+	},
+};
