+++ styles/MasteryIcons.tsx
